check-in code for a class to track the number of actions (think requests) per second in a sliding window capacity.
The purpose for this is to put on a server to track the amount of requests (or other stuff) that has happened globally within a second. If we're going to fire off an amount of something per second, it would make sense to have the server able to track and print this out.