check in start of code for idea formed for virtual users and wait intervals (inspired by locust and reading on other tools/performance testing intermittently throughout day) (IterationWaitStrategy, VirtualUser, SharedVirtualUserContext, VirtualUserManager)
Virtual users are users who do a set thing (more or less) a given or infinite amount of times, we'll want to have certain VirtualUsers active which do a set thing, and increase it to a given amount over time and see how the response changes. As we dial up the number of virtual users, we'll want to see how latency changes, we'll determine this after through reports, and when it comes to reporting, at the end of execution there is going to have to be context somewhere telling how many users were active at the time I imagine.