add new discovery about apt being prone to race conditions regarding file locks
When the apt install command has finished, stuff isn't necessarily done and back to normal. (A reoccurring theme). This option tacked onto the beginning (as we should probably start doing in automation on Debian & Ubuntu) makes it wait for the lock to be available instead of giving up immediately when it isn't.