add class for helping with executing playbooks against a host with a private key
I wanted to lay the way for executing a playbook against a single host that is referenced in memory, and where the key isn't initially written to disk, i.e. from some secrets manager in the not-so-distant future. The class will write take the private ssh RSA key, and the host of the system to execute given the playbooks against, and write the key to disk temporarily so that the ansible-playbook can be passed it internally. This solution may not meet all security requirements in the future since writing it to a TemporaryDirectory might be writing to disk, and writing to disk may leave it recoverable through undeletion, this is not a security concern here, for several reasons at this moment though.