add a main.tf terraform file for creating a droplet and provisioning it with ansible
The terraform file will create a droplet and generate an SSH key and store it in the state file. SSH on droplet won't be available immediately after creation, the playbook_executor.py script will wait for SSH availability and then run the script. The private key has to be written to a file, and this is handled by the script.