add start of experiment for replacing a component in pulumi when a property changes
This was isolated into its own experiment after it failed as part of a broader one to have a droplet destroyed when a dependency is replaced. An almost identical process for the usage of replace_on_changes worked earlier in dynamic.Resource/dynamic.ResourceProvider but not here. There is more to figure out for this, but enough is in place to experiment easier and demonstrate this issue here. It may be impossible to have Pulumi replace a ComponentResource with replace_on_changes or something.